requirements.txt
httpx==0.27.0
pydantic==2.9.2
python-dotenv==1.0.1
rich==13.9.2
numpy==1.26.4